data/requirements.txt
pandas
openpyxl
requests
beautifulsoup4
playwright
python-dotenv
supabase
